requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==22.1.0
certifi==2022.9.14
chardet==4.0.0
charset-normalizer==2.1.1
discord.py==1.7.3
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
prawcore==2.3.0
requests==2.28.1
typing_extensions==4.3.0
update-checker==0.18.0
urllib3==1.26.12
websocket-client==1.4.1
yarl==1.8.1
